Added issue-a-credential how-to for Python
