+++ doc/how-tos/issue-credential/python/step2.py
@@ -1,43 +1,3 @@
-"""
-This sample is extensions of "write_schema_and_cred_def.py"
-
-Shows how to issue a credential as a Trust Anchor which has created a Cred Definition
-for an existing Schema.
-
-After Trust Anchor has successfully created and stored a Cred Definition using Anonymous Credentials,
-Prover's wallet is created and opened, and used to generate Prover's Master Secret.
-After that, Trust Anchor generates Credential Offer for given Cred Definition, using Prover's DID
-Prover uses Credential Offer to create Credential Request
-Trust Anchor then uses Prover's Credential Request to issue a Credential.
-Finally, Prover stores Credential in its wallet.
-"""
-
-
-import asyncio
-import json
-import pprint
-
-from indy import pool, ledger, wallet, did, anoncreds
-from indy.error import IndyError
-
-
-pool_name = 'pool'
-genesis_file_path = '/home/vagrant/code/evernym/indy-sdk/cli/docker_pool_transactions_genesis'
-wallet_config = json.dumps({"id": "wallet"})
-wallet_credentials = json.dumps({"key": "wallet_key"})
-PROTOCOL_VERSION=2
-
-
-def print_log(value_color="", value_noncolor=""):
-    """set the colors for text."""
-    HEADER = '\033[92m'
-    ENDC = '\033[0m'
-    print(HEADER + value_color + ENDC + str(value_noncolor))
-
-
-async def issue_credential():
-    try:
-        await pool.set_protocol_version(2)
         # 1.
         print_log('\n1. Creates a new local pool ledger configuration that is used '
                   'later when connecting to ledger.\n')
@@ -136,73 +96,3 @@ async def issue_credential():
                                                                                cred_def_tag, cred_def_type, cred_def_config)
         print_log('Credential definition: ')
         pprint.pprint(json.loads(cred_def_json))
-
-        # 12.
-        print_log('\n12. Creating Prover wallet and opening it to get the handle\n')
-        prover_did = 'VsKV7grR1BUE29mG2Fm2kX'
-        prover_wallet_config = json.dumps({"id": "prover_wallet"})
-        prover_wallet_credentials = json.dumps({"key": "prover_wallet_key"})
-        await wallet.create_wallet(prover_wallet_config, prover_wallet_credentials)
-        prover_wallet_handle = await wallet.open_wallet(prover_wallet_config, prover_wallet_credentials)
-
-        # 13.
-        print_log('\n13. Prover is creating Master Secret\n')
-        master_secret_name = 'master_secret'
-        master_secret_id = await anoncreds.prover_create_master_secret(prover_wallet_handle, master_secret_name)
-
-        # 14.
-        print_log('\n14. Issuer (Trust Anchor) is creating a Credential Offer for Prover\n')
-        schema_json = json.dumps(schema)
-        cred_offer_json = await anoncreds.issuer_create_credential_offer(wallet_handle, cred_def_id)
-        print_log('Credential Offer: ')
-        pprint.pprint(json.loads(cred_offer_json))
-
-        # 15.
-        print_log('\n15. Prover creates Credential Request for the given credential offer\n')
-        (cred_req_json, cred_req_metadata_json) = await anoncreds.prover_create_credential_req(prover_wallet_handle, prover_did, cred_offer_json, cred_def_json, master_secret_id)
-        print_log('Credential Request: ')
-        pprint.pprint(json.loads(cred_req_json))
-
-        # 16.
-        print_log('\n16. Issuer (Trust Anchor) creates Credential for Credential Request\n')
-        cred_values_json = json.dumps({
-            'sex': ['male', '5944657099558967239210949258394887428692050081607692519917050011144233115103'],
-            'name': ['Alex', '1139481716457488690172217916278103335'],
-            'height': ['175', '175'],
-            'age': ['28', '28']
-        })
-        (cred_json, _, _) = await anoncreds.issuer_create_credential(wallet_handle, cred_offer_json, cred_req_json, cred_values_json, None, None)
-        print_log('Credential: ')
-        pprint.pprint(json.loads(cred_json))
-
-        # 17.
-        print_log('\n17. Prover processes and stores Credential\n')
-        await anoncreds.prover_store_credential(prover_wallet_handle, None, cred_req_metadata_json, cred_json, cred_def_json, None)
-
-        # 18.
-        print_log('\n18. Closing both wallet_handles and pool\n')
-        await wallet.close_wallet(wallet_handle)
-        await wallet.close_wallet(prover_wallet_handle)
-        await pool.close_pool_ledger(pool_handle)
-
-        # 19.
-        print_log('\n19. Deleting created wallet_handles\n')
-        await wallet.delete_wallet(wallet_config, wallet_credentials)
-        await wallet.delete_wallet(prover_wallet_config, prover_wallet_credentials)
-
-        # 20.
-        print_log('\n20. Deleting pool ledger config\n')
-        await pool.delete_pool_ledger_config(pool_name)
-
-    except IndyError as e:
-        print('Error occurred: %s' % e)
-
-
-def main():
-    loop = asyncio.get_event_loop()
-    loop.run_until_complete(issue_credential())
-    loop.close()
-
-
-if __name__ == '__main__':
-    main()
