+++ doc/how-tos/issue-credential/python/template.py
@@ -1,0 +1,58 @@
+"""
+This sample is extensions of "write_schema_and_cred_def.py"
+
+Shows how to issue a credential as a Trust Anchor which has created a Cred Definition
+for an existing Schema.
+
+After Trust Anchor has successfully created and stored a Cred Definition using Anonymous Credentials,
+Prover's wallet is created and opened, and used to generate Prover's Master Secret.
+After that, Trust Anchor generates Credential Offer for given Cred Definition, using Prover's DID
+Prover uses Credential Offer to create Credential Request
+Trust Anchor then uses Prover's Credential Request to issue a Credential.
+Finally, Prover stores Credential in its wallet.
+"""
+
+
+import asyncio
+import json
+import pprint
+
+from indy import pool, ledger, wallet, did, anoncreds
+from indy.error import IndyError
+
+
+pool_name = 'pool'
+genesis_file_path = '/home/vagrant/code/evernym/indy-sdk/cli/docker_pool_transactions_genesis'
+wallet_config = json.dumps({"id": "wallet"})
+wallet_credentials = json.dumps({"key": "wallet_key"})
+PROTOCOL_VERSION=2
+
+
+def print_log(value_color="", value_noncolor=""):
+    """set the colors for text."""
+    HEADER = '\033[92m'
+    ENDC = '\033[0m'
+    print(HEADER + value_color + ENDC + str(value_noncolor))
+
+
+async def issue_credential():
+    try:
+        await pool.set_protocol_version(2)
+        # Step 2 code goes here.
+
+        # Step 3 code goes here.
+
+        # Step 4 code goes here.
+
+    except IndyError as e:
+        print('Error occurred: %s' % e)
+
+
+def main():
+    loop = asyncio.get_event_loop()
+    loop.run_until_complete(issue_credential())
+    loop.close()
+
+
+if __name__ == '__main__':
+    main()
