+++ doc/how-tos/save-schema-and-cred-def/python/README.md
@@ -95,5 +95,5 @@ Save the updated version of `save_schema_and_cred_def.py``.
 
 ## More experiments
 
-You might try the ["Issue a Credential"](../../issue-credential/../not-yet-written.md)
+You might try the ["Issue a Credential"](../../issue-credential/python/README.md)
 how-to, which can be done in only one step once you complete this one.
